added view direction, raycast aiming around mouse

+++ Arena/Arena/Utils.cpp
@@ -11,7 +11,7 @@ Utils::~Utils(void)
 }
 
 
-double Utils::round(double f)
+int Utils::round(double f)
 {
 	return (f > 0.0) ? (f + 0.5) : (f - 0.5);
 }
